Extract composer editor shell
