+++ assets/js/composer.js
@@ -8,20 +8,20 @@ import {
   resolveSiteRepoConfig,
   parseYAML
 } from './yaml.js';
-import { t, getAvailableLangs, getLanguageLabel } from './i18n.js?v=press-system-v3.4.29';
-import { initSystemUpdates, getSystemUpdateSummaryEntries, getSystemUpdateCommitFiles, clearSystemUpdateState } from './system-updates.js?v=press-system-v3.4.29';
-import { initThemeManager, getThemeManagerSummaryEntries, getThemeManagerCommitFiles, clearThemeManagerState } from './theme-manager.js?v=press-system-v3.4.29';
-import { buildEditorContentTree, findEditorContentTreeNode, flattenEditorContentTree } from './editor-content-tree.js?v=press-system-v3.4.29';
+import { t, getAvailableLangs, getLanguageLabel } from './i18n.js?v=press-system-v3.4.30';
+import { initSystemUpdates, getSystemUpdateSummaryEntries, getSystemUpdateCommitFiles, clearSystemUpdateState } from './system-updates.js?v=press-system-v3.4.30';
+import { initThemeManager, getThemeManagerSummaryEntries, getThemeManagerCommitFiles, clearThemeManagerState } from './theme-manager.js?v=press-system-v3.4.30';
+import { buildEditorContentTree, findEditorContentTreeNode, flattenEditorContentTree } from './editor-content-tree.js?v=press-system-v3.4.30';
 import {
   decryptMarkdownDocument,
   encryptMarkdownDocument,
   parseEncryptedMarkdownEnvelope
-} from './encrypted-content.js?v=press-system-v3.4.29';
-import { createStagingRegistry } from './composer-staging.js?v=press-system-v3.4.29';
-import { createIndexPublishMetadataEnricher } from './composer-index-publish-metadata.js?v=press-system-v3.4.29';
-import { createContentCommitStagingProvider } from './composer-content-staging.js?v=press-system-v3.4.29';
-import { createSeoStagingProvider } from './composer-seo-staging.js?v=press-system-v3.4.29';
-import { createPostCommitStateApplier } from './composer-post-commit-state.js?v=press-system-v3.4.29';
+} from './encrypted-content.js?v=press-system-v3.4.30';
+import { createStagingRegistry } from './composer-staging.js?v=press-system-v3.4.30';
+import { createIndexPublishMetadataEnricher } from './composer-index-publish-metadata.js?v=press-system-v3.4.30';
+import { createContentCommitStagingProvider } from './composer-content-staging.js?v=press-system-v3.4.30';
+import { createSeoStagingProvider } from './composer-seo-staging.js?v=press-system-v3.4.30';
+import { createPostCommitStateApplier } from './composer-post-commit-state.js?v=press-system-v3.4.30';
 import {
   cloneIndexMetadataValue,
   computeIndexDiff,
@@ -35,7 +35,7 @@ import {
   prepareIndexState,
   prepareTabsState,
   safeString
-} from './composer-index-tabs-model.js?v=press-system-v3.4.29';
+} from './composer-index-tabs-model.js?v=press-system-v3.4.30';
 import {
   cloneSiteState,
   computeSiteDiff,
@@ -43,36 +43,33 @@ import {
   prepareSiteState,
   toSiteYaml,
   writeYamlValue
-} from './composer-site-model.js?v=press-system-v3.4.29';
+} from './composer-site-model.js?v=press-system-v3.4.30';
 import {
   createScopedStorageKey,
   resolveEditorStorageScope
-} from './editor-storage.js?v=press-system-v3.4.29';
-import { createScopedDraftStore } from './editor-drafts.js?v=press-system-v3.4.29';
-import { createEditorSessionStateStore } from './editor-session-state.js?v=press-system-v3.4.29';
+} from './editor-storage.js?v=press-system-v3.4.30';
+import { createScopedDraftStore } from './editor-drafts.js?v=press-system-v3.4.30';
+import { createEditorSessionStateStore } from './editor-session-state.js?v=press-system-v3.4.30';
 import {
   refreshSyncCommitPanelView,
   scheduleSyncCommitPanelRefreshView
-} from './composer-sync-panel.js?v=press-system-v3.4.29';
-import { createSyncOverlayController } from './composer-sync-overlay.js?v=press-system-v3.4.29';
-import {
-  animateEditorSystemPanelContent as animateSystemPanelContent,
-  showEditorSystemPanel as showComposerSystemPanel
-} from './composer-system-panel.js?v=press-system-v3.4.29';
-import { createPublishTransportSettingsUi } from './composer-publish-settings-ui.js?v=press-system-v3.4.29';
-import { createPublishSummaryRenderer } from './composer-publish-summary.js?v=press-system-v3.4.29';
-import { createComposerPublishFlow } from './composer-publish-flow.js?v=press-system-v3.4.29';
-import { createComposerDiffUi } from './composer-diff-ui.js?v=press-system-v3.4.29';
-import { createComposerOrderDiffUi } from './composer-order-diff-ui.js?v=press-system-v3.4.29';
-import { createComposerIndexTabsUi } from './composer-index-tabs-ui.js?v=press-system-v3.4.29';
-import { createComposerSiteSettingsUi } from './composer-site-settings-ui.js?v=press-system-v3.4.29';
-import { createComposerMarkdownAssetManager } from './composer-markdown-assets.js?v=press-system-v3.4.29';
-import { createEditorFileTreeUi } from './editor-file-tree-ui.js?v=press-system-v3.4.29';
-import { createEditorStructurePanelUi } from './editor-structure-panel-ui.js?v=press-system-v3.4.29';
+} from './composer-sync-panel.js?v=press-system-v3.4.30';
+import { createSyncOverlayController } from './composer-sync-overlay.js?v=press-system-v3.4.30';
+import { createPublishTransportSettingsUi } from './composer-publish-settings-ui.js?v=press-system-v3.4.30';
+import { createPublishSummaryRenderer } from './composer-publish-summary.js?v=press-system-v3.4.30';
+import { createComposerPublishFlow } from './composer-publish-flow.js?v=press-system-v3.4.30';
+import { createComposerDiffUi } from './composer-diff-ui.js?v=press-system-v3.4.30';
+import { createComposerOrderDiffUi } from './composer-order-diff-ui.js?v=press-system-v3.4.30';
+import { createComposerIndexTabsUi } from './composer-index-tabs-ui.js?v=press-system-v3.4.30';
+import { createComposerSiteSettingsUi } from './composer-site-settings-ui.js?v=press-system-v3.4.30';
+import { createComposerMarkdownAssetManager } from './composer-markdown-assets.js?v=press-system-v3.4.30';
+import { createComposerEditorShell } from './composer-editor-shell.js?v=press-system-v3.4.30';
+import { createEditorFileTreeUi } from './editor-file-tree-ui.js?v=press-system-v3.4.30';
+import { createEditorStructurePanelUi } from './editor-structure-panel-ui.js?v=press-system-v3.4.30';
 import {
   CONNECT_PUBLISH_PRESETS,
   createPublishSettingsStore
-} from './publish/settings-store.js?v=press-system-v3.4.29';
+} from './publish/settings-store.js?v=press-system-v3.4.30';
 
 // Utility helpers
 const $ = (s, r = document) => r.querySelector(s);
@@ -135,7 +132,6 @@ const LS_KEYS = {
   systemTreeExpanded: 'press_editor_system_tree_expanded'
 };
 const EDITOR_STATE_VERSION = 3;
-const EDITOR_SCROLL_SAVE_DELAY = 120;
 
 const EDITOR_STORAGE_SCOPE = (() => {
   try { return resolveEditorStorageScope(window.location); }
@@ -436,18 +432,43 @@ try {
     expandedEditorTreeNodeIds.add('system');
   }
 } catch (_) {}
-let activeEditorOverlayMode = null;
-let editorOverlayReturnFocus = null;
-let editorOverlayEscapeBound = false;
-let editorRailResizeBound = false;
-let editorMobileRailBound = false;
-let editorStatePersistTimer = 0;
-let editorStateScrollBound = false;
-let editorContentScrollByKey = {};
-const EDITOR_RAIL_WIDTH_KEY = 'press_editor_rail_width';
-const EDITOR_RAIL_DEFAULT_WIDTH = 340;
-const EDITOR_RAIL_MIN_WIDTH = 280;
-const EDITOR_RAIL_MAX_WIDTH = 520;
+const editorShell = createComposerEditorShell({
+  documentRef: document,
+  windowRef: window,
+  editorSessionStateStore,
+  expandedEditorTreeNodeIds,
+  treeText,
+  getCurrentMode: () => currentMode,
+  getDynamicEditorTabs: () => dynamicEditorTabs,
+  isDynamicMode,
+  normalizeRelPath,
+  getAllowEditorStatePersist: () => allowEditorStatePersist,
+  persistDynamicEditorState,
+  getActiveComposerFile,
+  applyComposerFile,
+  refreshSyncCommitPanel,
+  applyMode
+});
+const {
+  mountEditorSystemPanels,
+  setEditorSystemPanelVisible,
+  showEditorSystemPanel,
+  initEditorOverlay,
+  openEditorOverlay,
+  initEditorRailResize,
+  initMobileEditorRail,
+  closeEditorRailDrawer,
+  bindEditorStatePersistenceListeners,
+  persistSystemTreeExpandedState,
+  scheduleEditorStatePersist,
+  captureEditorContentScroll,
+  restoreEditorContentScrollForMode,
+  scrollEditorContentToTop,
+  getEditorRailScrollTop,
+  setEditorRailScrollTop,
+  setEditorContentScrollByKey,
+  getEditorContentScrollSnapshot
+} = editorShell;
 
 function getDynamicTabsContainer() {
   try {
@@ -4622,537 +4643,6 @@ function deriveDynamicTabIdentity(path, options = {}) {
   };
 }
 
-function getEditorContentPane() {
-  try { return document.getElementById('editorContentPane'); }
-  catch (_) { return null; }
-}
-
-function scrollEditorContentToTop(behavior = 'smooth') {
-  const pane = getEditorContentScrollElement(currentMode) || getEditorContentPane();
-  if (pane && typeof pane.scrollTo === 'function') {
-    try {
-      pane.scrollTo({ top: 0, behavior });
-      captureEditorContentScroll(currentMode);
-      return;
-    } catch (_) {
-      try {
-        pane.scrollTop = 0;
-        captureEditorContentScroll(currentMode);
-        return;
-      } catch (__) {}
-    }
-  }
-  if (typeof window !== 'undefined' && typeof window.scrollTo === 'function') {
-    try { window.scrollTo({ top: 0, behavior }); }
-    catch (_) { try { window.scrollTo(0, 0); } catch (__) {} }
-  }
-}
-
-function persistSystemTreeExpandedState() {
-  editorSessionStateStore.writeLegacySystemTreeExpanded(expandedEditorTreeNodeIds.has('system'));
-}
-
-function normalizeEditorScrollTop(value) {
-  const numeric = Number(value);
-  if (!Number.isFinite(numeric)) return 0;
-  return Math.max(0, Math.round(numeric));
-}
-
-function normalizeEditorScrollMap(value) {
-  const out = {};
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return out;
-  Object.entries(value).forEach(([key, top]) => {
-    const normalizedKey = String(key || '').trim();
-    if (!normalizedKey) return;
-    out[normalizedKey] = normalizeEditorScrollTop(top);
-  });
-  return out;
-}
-
-function getEditorRailScrollElement() {
-  try { return document.querySelector('.editor-rail-tree-scroll'); }
-  catch (_) { return null; }
-}
-
-function getEditorRailScrollTop() {
-  const rail = getEditorRailScrollElement();
-  try { return rail ? normalizeEditorScrollTop(rail.scrollTop || 0) : 0; }
-  catch (_) { return 0; }
-}
-
-function setEditorRailScrollTop(top) {
-  const rail = getEditorRailScrollElement();
-  if (!rail) return;
-  try { rail.scrollTop = normalizeEditorScrollTop(top); } catch (_) {}
-}
-
-function getEditorContentScrollKey(mode = currentMode) {
-  if (mode && isDynamicMode(mode)) {
-    const tab = dynamicEditorTabs.get(mode);
-    const path = tab && tab.path ? normalizeRelPath(tab.path) : '';
-    return path ? `markdown:${path}` : 'markdown';
-  }
-  if (mode === 'composer') return 'composer';
-  if (mode === 'themes') return 'themes';
-  if (mode === 'updates') return 'updates';
-  if (mode === 'sync') return 'sync';
-  return 'structure';
-}
-
-function getEditorContentScrollElement(mode = currentMode) {
-  try {
-    if (mode === 'composer') {
-      const siteViewport = document.querySelector('#composerSite .cs-viewport');
-      if (siteViewport && siteViewport.getClientRects && siteViewport.getClientRects().length) {
-        return siteViewport;
-      }
-    }
-  } catch (_) {}
-  return getEditorContentPane();
-}
-
-function getEditorContentScrollTop(mode = currentMode) {
-  const scroller = getEditorContentScrollElement(mode);
-  try { return scroller ? normalizeEditorScrollTop(scroller.scrollTop || 0) : 0; }
-  catch (_) { return 0; }
-}
-
-function setEditorContentScrollTopForMode(mode, top) {
-  const scroller = getEditorContentScrollElement(mode);
-  if (!scroller) return false;
-  try {
-    scroller.scrollTop = normalizeEditorScrollTop(top);
-    return true;
-  } catch (_) {
-    return false;
-  }
-}
-
-function captureEditorContentScroll(mode = currentMode) {
-  const key = getEditorContentScrollKey(mode);
-  if (!key) return;
-  editorContentScrollByKey[key] = getEditorContentScrollTop(mode);
-}
-
-function restoreEditorContentScrollForMode(mode = currentMode) {
-  const key = getEditorContentScrollKey(mode);
-  if (!key || !Object.prototype.hasOwnProperty.call(editorContentScrollByKey, key)) return;
-  const top = editorContentScrollByKey[key];
-  const apply = () => setEditorContentScrollTopForMode(mode, top);
-  try {
-    requestAnimationFrame(() => requestAnimationFrame(apply));
-  } catch (_) {
-    setTimeout(apply, 0);
-  }
-}
-
-function scheduleEditorStatePersist() {
-  if (!allowEditorStatePersist) return;
-  try {
-    if (editorStatePersistTimer) window.clearTimeout(editorStatePersistTimer);
-    editorStatePersistTimer = window.setTimeout(() => {
-      editorStatePersistTimer = 0;
-      persistDynamicEditorState();
-    }, EDITOR_SCROLL_SAVE_DELAY);
-  } catch (_) {
-    persistDynamicEditorState();
-  }
-}
-
-function bindEditorStatePersistenceListeners() {
-  if (editorStateScrollBound) return;
-  editorStateScrollBound = true;
-  const onScroll = (event) => {
-    const target = event && event.target;
-    try {
-      if (target === getEditorRailScrollElement()) {
-        scheduleEditorStatePersist();
-        return;
-      }
-      if (target === getEditorContentPane() || target === getEditorContentScrollElement(currentMode)) {
-        captureEditorContentScroll(currentMode);
-        scheduleEditorStatePersist();
-      }
-    } catch (_) {}
-  };
-  try { document.addEventListener('scroll', onScroll, true); } catch (_) {}
-  try { window.addEventListener('pagehide', () => persistDynamicEditorState()); } catch (_) {}
-  try {
-    document.addEventListener('visibilitychange', () => {
-      if (document.visibilityState === 'hidden') persistDynamicEditorState();
-    });
-  } catch (_) {}
-}
-
-function mountEditorSystemPanels() {
-  const body = document.getElementById('editorSystemBody');
-  if (!body || body.__pressSystemPanelsMounted) return;
-  body.__pressSystemPanelsMounted = true;
-  const composerPanel = document.getElementById('mode-composer');
-  const themesPanel = document.getElementById('mode-themes');
-  const updatesPanel = document.getElementById('mode-updates');
-  let syncPanel = document.getElementById('mode-sync');
-  if (!syncPanel) {
-    syncPanel = document.createElement('div');
-    syncPanel.id = 'mode-sync';
-    syncPanel.className = 'sync-layout editor-overlay-panel';
-    syncPanel.hidden = true;
-    syncPanel.setAttribute('aria-hidden', 'true');
-  }
-  if (composerPanel) body.appendChild(composerPanel);
-  if (themesPanel) body.appendChild(themesPanel);
-  if (updatesPanel) body.appendChild(updatesPanel);
-  if (syncPanel) body.appendChild(syncPanel);
-}
-
-function setEditorSystemPanelVisible(visible) {
-  const panel = document.getElementById('editorSystemPanel');
-  if (!panel) return;
-  if (visible) {
-    panel.removeAttribute('hidden');
-    panel.removeAttribute('aria-hidden');
-  } else {
-    panel.setAttribute('hidden', '');
-    panel.setAttribute('aria-hidden', 'true');
-    panel.classList.remove('is-content-entering');
-  }
-}
-
-function animateEditorSystemPanelContent() {
-  animateSystemPanelContent({ windowRef: window, documentRef: document });
-}
-
-function showEditorSystemPanel(mode) {
-  return showComposerSystemPanel(mode, {
-    documentRef: document,
-    treeText,
-    mountEditorSystemPanels,
-    setEditorSystemPanelVisible,
-    getActiveComposerFile,
-    applyComposerFile,
-    resetSiteSettingsNavOnOpen,
-    refreshSyncCommitPanel,
-    animatePanel: animateEditorSystemPanelContent
-  });
-}
-
-function getEditorOverlayTitle(mode) {
-  if (mode === 'sync') return treeText('sync', 'Publish');
-  if (mode === 'updates') return t('editor.systemUpdates.tabLabel');
-  if (mode === 'themes') return treeText('themes', 'Themes');
-  return t('editor.modes.composer');
-}
-
-function syncEditorOverlayUi() {
-  const layer = document.getElementById('editorModalLayer');
-  const dialog = document.querySelector('.editor-modal-dialog');
-  const title = document.getElementById('editorModalTitle');
-  const modalBody = document.querySelector('.editor-modal-body');
-  const hasOverlay = false;
-
-  if (layer) {
-    layer.hidden = true;
-    layer.setAttribute('aria-hidden', 'true');
-  }
-  if (title) title.textContent = '';
-  if (dialog) dialog.removeAttribute('aria-label');
-  if (modalBody) {
-    modalBody.classList.remove('is-composer-overlay');
-    modalBody.classList.remove('is-updates-overlay');
-  }
-
-  try {
-    document.body.classList.toggle('press-editor-modal-open', hasOverlay);
-  } catch (_) {}
-}
-
-function focusEditorOverlay() {
-  const dialog = document.querySelector('.editor-modal-dialog');
-  if (!dialog) return;
-  const selector = 'button:not([disabled]), [href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])';
-  let target = null;
-  try {
-    target = dialog.querySelector(selector);
-  } catch (_) {
-    target = null;
-  }
-  const focusTarget = target || dialog;
-  try { focusTarget.focus({ preventScroll: true }); }
-  catch (_) { try { focusTarget.focus(); } catch (__) {} }
-}
-
-function resetSiteSettingsNavOnOpen() {
-  const root = document.getElementById('composerSite');
-  if (!root) return;
-  try {
-    const viewport = root.querySelector('.cs-viewport');
-    if (viewport) viewport.scrollTop = 0;
-  } catch (_) {}
-  try {
-    const modalBody = typeof root.closest === 'function' ? root.closest('.editor-modal-body') : null;
-    if (modalBody) modalBody.scrollTop = 0;
-  } catch (_) {}
-  const firstSectionId = (() => {
-    try { return String(root.__pressSiteFirstSectionId || '').trim(); }
-    catch (_) { return ''; }
-  })();
-  const setActive = (() => {
-    try { return typeof root.__pressSiteNavSetActive === 'function' ? root.__pressSiteNavSetActive : null; }
-    catch (_) { return null; }
-  })();
-  if (!firstSectionId || !setActive) return;
-  const activateFirst = () => {
-    try {
-      setActive(firstSectionId, {
-        focusPanel: false,
-        scrollViewport: false,
-        skipScrollLock: true
-      });
-    } catch (_) {}
-  };
-  activateFirst();
-  try {
-    requestAnimationFrame(() => requestAnimationFrame(activateFirst));
-  } catch (_) {
-    activateFirst();
-  }
-}
-
-function openEditorOverlay(mode, trigger = null) {
-  const nextMode = mode === 'updates' ? 'updates' : (mode === 'themes' ? 'themes' : mode === 'composer' ? 'composer' : null);
-  if (!nextMode) return;
-  editorOverlayReturnFocus = trigger && typeof trigger.focus === 'function' ? trigger : document.activeElement;
-  activeEditorOverlayMode = null;
-  applyMode(nextMode, { trigger });
-}
-
-function closeEditorOverlay() {
-  if (!activeEditorOverlayMode) return;
-  activeEditorOverlayMode = null;
-  syncEditorOverlayUi();
-  const restore = editorOverlayReturnFocus;
-  editorOverlayReturnFocus = null;
-  if (restore && typeof restore.focus === 'function') {
-    try { restore.focus({ preventScroll: true }); }
-    catch (_) { try { restore.focus(); } catch (__) {} }
-  }
-}
-
-function initEditorOverlay() {
-  const layer = document.getElementById('editorModalLayer');
-  if (!layer || layer.__editorOverlayBound) return;
-  layer.__editorOverlayBound = true;
-  layer.addEventListener('click', (event) => {
-    const target = event.target;
-    if (target && target.closest && target.closest('[data-editor-modal-close]')) closeEditorOverlay();
-  });
-  const closeBtn = document.getElementById('editorModalClose');
-  if (closeBtn && !closeBtn.__editorOverlayCloseBound) {
-    closeBtn.__editorOverlayCloseBound = true;
-    closeBtn.addEventListener('click', closeEditorOverlay);
-  }
-  if (!editorOverlayEscapeBound) {
-    editorOverlayEscapeBound = true;
-    document.addEventListener('keydown', (event) => {
-      if (!activeEditorOverlayMode) return;
-      if (event.key === 'Escape') {
-        event.preventDefault();
-        closeEditorOverlay();
-        return;
-      }
-      if (event.key !== 'Tab') return;
-      const dialog = document.querySelector('.editor-modal-dialog');
-      if (!dialog) return;
-      let focusables = [];
-      try {
-        focusables = Array.from(dialog.querySelectorAll('button:not([disabled]), [href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])'))
-          .filter((el) => el && el.offsetParent !== null);
-      } catch (_) {
-        focusables = [];
-      }
-      if (!focusables.length) {
-        event.preventDefault();
-        try { dialog.focus({ preventScroll: true }); } catch (_) { try { dialog.focus(); } catch (__) {} }
-        return;
-      }
-      const first = focusables[0];
-      const last = focusables[focusables.length - 1];
-      if (event.shiftKey && document.activeElement === first) {
-        event.preventDefault();
-        try { last.focus({ preventScroll: true }); } catch (_) { last.focus(); }
-      } else if (!event.shiftKey && document.activeElement === last) {
-        event.preventDefault();
-        try { first.focus({ preventScroll: true }); } catch (_) { first.focus(); }
-      }
-    });
-  }
-  syncEditorOverlayUi();
-}
-
-function computeEditorRailMaxWidth() {
-  let viewportLimit = EDITOR_RAIL_MAX_WIDTH;
-  try {
-    if (typeof window !== 'undefined' && Number.isFinite(window.innerWidth)) {
-      viewportLimit = Math.min(EDITOR_RAIL_MAX_WIDTH, window.innerWidth * 0.46);
-    }
-  } catch (_) {}
-  return Math.max(EDITOR_RAIL_MIN_WIDTH, viewportLimit);
-}
-
-function clampEditorRailWidth(value) {
-  const numeric = Number(value);
-  const fallback = EDITOR_RAIL_DEFAULT_WIDTH;
-  const width = Number.isFinite(numeric) ? numeric : fallback;
-  return Math.max(EDITOR_RAIL_MIN_WIDTH, Math.min(computeEditorRailMaxWidth(), width));
-}
-
-function setEditorRailWidth(value, options = {}) {
-  const width = clampEditorRailWidth(value);
-  try {
-    document.documentElement.style.setProperty('--editor-rail-width', `${Math.round(width)}px`);
-  } catch (_) {}
-  const resizer = document.getElementById('editorRailResizer');
-  if (resizer) {
-    resizer.setAttribute('aria-valuemin', String(EDITOR_RAIL_MIN_WIDTH));
-    resizer.setAttribute('aria-valuemax', String(Math.round(computeEditorRailMaxWidth())));
-    resizer.setAttribute('aria-valuenow', String(Math.round(width)));
-  }
-  if (options.persist) {
-    editorSessionStateStore.writeUnscopedNumber(EDITOR_RAIL_WIDTH_KEY, width);
-  }
-  return width;
-}
-
-function initEditorRailResize() {
-  if (editorRailResizeBound) return;
-  const resizer = document.getElementById('editorRailResizer');
-  const shell = document.getElementById('editorAppShell');
-  if (!resizer || !shell) return;
-  editorRailResizeBound = true;
-
-  let stored = EDITOR_RAIL_DEFAULT_WIDTH;
-  stored = editorSessionStateStore.readUnscopedNumber(EDITOR_RAIL_WIDTH_KEY, EDITOR_RAIL_DEFAULT_WIDTH);
-  setEditorRailWidth(stored, { persist: false });
-
-  const isMobile = () => {
-    try {
-      return !!(window.matchMedia && window.matchMedia('(max-width: 820px)').matches);
-    } catch (_) {
-      return false;
-    }
-  };
-
-  let dragState = null;
-  const finishDrag = () => {
-    if (!dragState) return;
-    const width = dragState.width;
-    dragState = null;
-    shell.classList.remove('is-resizing-rail');
-    try { document.body.style.removeProperty('cursor'); } catch (_) {}
-    setEditorRailWidth(width, { persist: true });
-    document.removeEventListener('pointermove', onMove);
-    document.removeEventListener('pointerup', finishDrag);
-    document.removeEventListener('pointercancel', finishDrag);
-  };
-  const onMove = (event) => {
-    if (!dragState || isMobile()) return;
-    const delta = Number(event.clientX) - dragState.startX;
-    dragState.width = setEditorRailWidth(dragState.startWidth + delta, { persist: false });
-  };
-
-  resizer.addEventListener('pointerdown', (event) => {
-    if (isMobile()) return;
-    event.preventDefault();
-    dragState = {
-      startX: Number(event.clientX) || 0,
-      startWidth: setEditorRailWidth(resizer.getAttribute('aria-valuenow') || EDITOR_RAIL_DEFAULT_WIDTH, { persist: false }),
-      width: EDITOR_RAIL_DEFAULT_WIDTH,
-    };
-    dragState.width = dragState.startWidth;
-    shell.classList.add('is-resizing-rail');
-    try { document.body.style.cursor = 'col-resize'; } catch (_) {}
-    document.addEventListener('pointermove', onMove);
-    document.addEventListener('pointerup', finishDrag);
-    document.addEventListener('pointercancel', finishDrag);
-  });
-
-  resizer.addEventListener('keydown', (event) => {
-    if (isMobile()) return;
-    let delta = 0;
-    if (event.key === 'ArrowLeft') delta = -16;
-    else if (event.key === 'ArrowRight') delta = 16;
-    else if (event.key === 'Home') {
-      event.preventDefault();
-      setEditorRailWidth(EDITOR_RAIL_DEFAULT_WIDTH, { persist: true });
-      return;
-    } else {
-      return;
-    }
-    event.preventDefault();
-    const current = Number(resizer.getAttribute('aria-valuenow')) || EDITOR_RAIL_DEFAULT_WIDTH;
-    setEditorRailWidth(current + delta, { persist: true });
-  });
-
-  window.addEventListener('resize', () => {
-    const current = Number(resizer.getAttribute('aria-valuenow')) || EDITOR_RAIL_DEFAULT_WIDTH;
-    setEditorRailWidth(current, { persist: false });
-  });
-}
-
-function isEditorMobileRailLayout() {
-  try {
-    return !!(window.matchMedia && window.matchMedia('(max-width: 820px)').matches);
-  } catch (_) {
-    return false;
-  }
-}
-
-function getEditorRailToggles() {
-  return Array.from(document.querySelectorAll('[data-editor-rail-toggle]'));
-}
-
-function setEditorRailOpen(open) {
-  const shell = document.getElementById('editorAppShell');
-  const toggles = getEditorRailToggles();
-  const scrim = document.getElementById('editorRailScrim');
-  if (!shell) return;
-  const shouldOpen = !!open && isEditorMobileRailLayout();
-  shell.classList.toggle('is-rail-open', shouldOpen);
-  toggles.forEach((toggle) => {
-    toggle.setAttribute('aria-expanded', shouldOpen ? 'true' : 'false');
-  });
-  if (scrim) {
-    scrim.hidden = !shouldOpen;
-    scrim.setAttribute('aria-hidden', shouldOpen ? 'false' : 'true');
-  }
-}
-
-function closeEditorRailDrawer() {
-  setEditorRailOpen(false);
-}
-
-function initMobileEditorRail() {
-  if (editorMobileRailBound) return;
-  const toggles = getEditorRailToggles();
-  const scrim = document.getElementById('editorRailScrim');
-  if (!toggles.length) return;
-  editorMobileRailBound = true;
-  toggles.forEach((toggle) => {
-    toggle.addEventListener('click', () => {
-      const shell = document.getElementById('editorAppShell');
-      const isOpen = !!(shell && shell.classList.contains('is-rail-open'));
-      setEditorRailOpen(!isOpen);
-    });
-  });
-  if (scrim) scrim.addEventListener('click', closeEditorRailDrawer);
-  document.addEventListener('keydown', (event) => {
-    if (event.key === 'Escape') closeEditorRailDrawer();
-  });
-  window.addEventListener('resize', () => {
-    if (!isEditorMobileRailLayout()) closeEditorRailDrawer();
-  });
-}
-
 function persistDynamicEditorState() {
   if (!allowEditorStatePersist) return;
   try {
@@ -5183,7 +4673,7 @@ function persistDynamicEditorState() {
       open,
       expandedNodeIds: Array.from(expandedEditorTreeNodeIds).filter(Boolean),
       railScrollTop: getEditorRailScrollTop(),
-      contentScrollByKey: { ...editorContentScrollByKey },
+      contentScrollByKey: getEditorContentScrollSnapshot(),
       updatedAt: Date.now()
     };
     if (currentMode && isDynamicMode(currentMode)) {
@@ -5199,7 +4689,7 @@ function restoreDynamicEditorState() {
   if (!data || typeof data !== 'object') return false;
 
   const isV3 = data.v === EDITOR_STATE_VERSION;
-  editorContentScrollByKey = normalizeEditorScrollMap(data.contentScrollByKey);
+  setEditorContentScrollByKey(data.contentScrollByKey);
 
   const open = Array.isArray(data.open) ? data.open : [];
   const seen = new Set();
