+++ scripts/test-system-release-package.sh
@@ -101,6 +101,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-markdown-assets.js" "$
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-editor-shell.js" "${entries_file}"; then
+  echo "expected package to include composer editor shell code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-file-tree-ui.js" "${entries_file}"; then
   echo "expected package to include editor file tree UI code" >&2
   exit 1
